requirements.txt
Django==5.2.3
django-ninja==1.3.0
psycopg2-binary==2.9.10
django-extensions==4.1
osmnx>=1.9.4
geopandas>=1.0.1
shapely>=2.0.6
folium>=0.17.0
matplotlib>=3.9.2
requests>=2.32.3
# Climate data packages
python-dotenv>=1.0.0
aiohttp>=3.9.0
